requirements.txt
streamlit
tensorflow-cpu
opencv-python
numpy
pillow
